Implements basic dynamic xml functionality
For both source and sink xml blocks, gain fields change according
to the device. Supported devices include uhd b200, limesdrmini,
plutosdr. Other devices are still functional with the use of
overall gain as it was. Also for the aformentioned devices,
checks are made to the gain values to be inside the specified
range. Also the number of channels are being checked according
to the device, with 2 being the maximum number of channels
supported right now.