Fixes bug in overall gain
When gain mode is set to auto the callback to set gain is not called
for supported devices. If a variable was set on a specific gain and
its value changed, the corresponding callback was called and it
seemed like the Overall Gain was working properly.
This bug was created from commit
f382e9c7f9b41da993b91753fda4a5e383e5754e and affected all supported
devices.