Fix RX stream
This commit fixes an issue first encountered using the
lower sampling rates of the RTL SDR.

However, the problem was deeper and it seems that the RX
stream was not properly handled. The improved way tells the
GNU Radio scheduler to handle items in multiple of 1024.
Due to internal GNU Radio stuff, this will eventually
lead to smaller circular buffers improvuing the cache
performance.

Also instead of recursive calls of the thw work() method,
the readStream() is called in a loop trying always to read,
except in case of a severe error. This will eliminate
the quite slow response in case of concecutive timeouts,
leading to a quite deep call stack.