Fix TX issue with PlutoSDR
The module of PlutoSDR returns 0 for the MTU of the TX stream.
So if the MTU is zero, the sink now uses the default of GNU Radio
maximum number of items. This is ok, because normally the
writeStream() of each module, consumes only the minimum
number of items that can copy to the internal buffers.