Source: Do not set the output multiple
The set_output_multiple() at the source block had a hardcoded
value of 1024. Some devices (like the USRP N-Series) may have
an MTU that is smaller than 1024. In this case the scheduler
hangs and the flowgraph cannot execute. As in the work() method
we handle items only in a per MTU basis, the set_output_multiple()
is not needed and we can clamp the available items in the
work() by using the set_max_noutput_items(MTU).

Signed-off-by: Manolis Surligas <surligas@gmail.com>